feat(java/*): add update and delete api for item, order and payment

+++ src/main/java/com/sjsu/project/cmpe202/controller/ItemController.java
@@ -4,6 +4,7 @@ import com.sjsu.project.cmpe202.model.Item;
 import com.sjsu.project.cmpe202.repository.ItemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -23,8 +24,8 @@ public class ItemController {
             value = "/item/get_all_items",
             method = RequestMethod.GET,
             consumes = "application/json")
-    public List<Item> getAllItems() {
-        return itemRepository.findAll();
+    public ResponseEntity<List<Item>> getAllItems() {
+        return new ResponseEntity<>(itemRepository.findAll(), HttpStatus.OK);
     }
 
     @CrossOrigin(origins = "http://localhost:8080")
@@ -32,14 +33,18 @@ public class ItemController {
             value = "/item/add_new_item",
             method = RequestMethod.POST,
             consumes = "application/json")
-    public HttpStatus addNewItem(Map<String, String> item) {
+    public ResponseEntity<String> addNewItem(Map<String, String> item) {
+        if (!item.containsKey("price"))
+            return new ResponseEntity<>("Missing parameter price", HttpStatus.BAD_REQUEST);
+        if (!item.containsKey("product"))
+            return new ResponseEntity<>("Missing parameter product", HttpStatus.BAD_REQUEST);
         Item newItem = new Item();
         newItem.setImageUrl(item.get("image_url"));
         Double price = Double.parseDouble(item.get("price"));
         newItem.setPrice(price);
         newItem.setProduct(item.get("product"));
         itemRepository.save(newItem);
-        return HttpStatus.OK;
+        return new ResponseEntity<>("Add New Item successfully", HttpStatus.OK);
     }
 
     @CrossOrigin(origins = "http://localhost:8080")
@@ -47,8 +52,10 @@ public class ItemController {
             value = "/item/delete_item_by_item_id",
             method = RequestMethod.DELETE,
             consumes = "application/json")
-    public HttpStatus deleteItemByItemId(Map<String, Integer> item) {
+    public ResponseEntity<String> deleteItemByItemId(Map<String, Integer> item) {
+        if (!item.containsKey("item_id"))
+            return new ResponseEntity<>("Missing parameter item_id", HttpStatus.BAD_REQUEST);
         itemRepository.deleteById(item.get("item_id"));
-        return HttpStatus.OK;
+        return new ResponseEntity<>("Delete Item successfully", HttpStatus.OK);
     }
 }
